Offcpu profiling
This new module calls the callback if a process has been scheduled out
for more than the threshold (configurable at runtime in /sys). To limit
the noise, we skip the kernel threads. The default threshold is 5 seconds.

The current callback emits a trace event and also dumps the kernel
stack in dmesg in a format compatible with stackcollapse.pl
https://github.com/brendangregg/FlameGraph

To use it: just load tracker.ko and offcpu.ko

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>