Allow user-space programs to wait on wakeup latency events
A new /proc/wake_latency file is created, when a latency event is
generated the readers waiting to read/poll on this file are woken up so
they can take an action when the high latency occurs.
An example use-case is provided in watch-wake-latency-event.sh where we
record an LTTng snapshot everytime a high latency is detected.
For now a hard-coded rate-limiter of 1 wakeup/second is in place, we'll
see if we keep it there.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>