New timeout architecture
Instead of creating one timer for each event (which does not scale),
create a list of pending events and use a thread to check the expiration
dates periodically. We loose a little in precision (granularity of the
thread frequency), but it is much more scalable.
To avoid creating a bottleneck, this patch needed a wait-free concurrent
queue which is not available in the kernel, so it comes with a port of
URCU wfcqueue.
The ownership of the events is now dependant on a refcount depending if
we need to track a timeout or not.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>