Take the stack in a reliable way
Instead of taking the stack when the timeout is reached, wait for the
next time the process is going to get scheduled in to capture it.
Limitation: for now we use the same HT than the registered processes
one, so only registered processes have their stack dumped. We'll see if
we create a new HT for that or add a "registered flag" only set when a
process is registered.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>