Fix: use the get/put event to take the stackdump
Much better integration with the tracker mechanism, a lot less memory
used.
Side-effect: for now we record the stack every time the process waiting
gets some CPU time until the syscall is completed, so we see its
progression in the kernel once it has been waiting for more than the
timeout parameter. Maybe it is too much, we'll see if we keep it that
way.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>