Implement our own call_rcu for kernel < 3.17
The default call_rcu on 3.13 causes sched_wakeups that we instrument in
the offcpu module, so it causes deadlocks.
This patch replaces the default call_rcu with a custom one implemented
as a delayed work which is done in a separate context so it avoids the
deadlock.
This allows us to use the lockless linked-list on 3.13 as well.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>