Remove rhashtable and annoying compile warnings
rhashtable does not add anything compared to the kernel-ported URCU HT.
URCU HT has a lock-free insertion which is not present in rhashtable and
rhashtable is only available in kernel >= 3.17 whereas URCU HT is
shipped with this module.
In terms of performance, the 2 HT are similar for lookup and
single-threaded insertion.
Finally, the URCU HT can be resized in a separate process (not ported
yet), whereas rhashtable needs to be resized during an insert/remove
which poses problem for our use-case where we need to avoid any memory
allocation during insert/remove.

Removing it simplifies the amount of #ifdef in the code and increases
the readability a bit.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>