Allocate private data
The users can now preallocate memory for each event and access it with
get_event and in the callback, so it is possible to store data inside
the event and get it back later. The refcount guarantees that the memory
is allocated and not reset, but does not provide any kind of exclusive
access, so the user needs to make sure the same event cannot be modified
at multiple places at the same time or use that data space to store a
lock.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>