copy the text payload for each event + refactoring
Having a single event that contains the text data about the callchain
does not work if we want to handle multiple branches originating from
the same interrupt (ex: multiple sched_wakeup from a softirq).

Good refactoring to have a single function handle all the state
transitions, much clearer.

The breakdown does not match the total anymore, need to fix that.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>