NUMA-aware allocator
Allocating memory per-CPU with a shared pool, has the side-effect of
requiring a lot of useless memory for unbalanced workloads. Also, if the
machine has a NUMA topology, it creates a lot of remote accesses.

This new approach creates a shared pool for each NUMA node and the CPUs
keep only a local cache of 128 events to limit false sharing. If a CPU
needs memory and its NUMA pool is empty, the allocation fails. When
putting back an event that belongs to a remote NUMA node, we return it
to its original pool.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>