Add event pointer parameter to latency_tracker_event_out
If the caller already has a valid reference the event it wants to end,
it can pass it directly to latency_tracker_event_out instead of
requiring another lookup which might find a different event.

Also, remove the lookup of the event_in from event_transition, now the
caller performs this lookup before entering the transition and releases
the reference afterwards.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>