Add tracking_on debugfs entry
This allows to enable/disable a tracker without loading/removing the
tracker modules. A callback is provided for the trackers if they want to
be informed of value changes in the tracking_on file. The callback is
called after the tracking_on value has been changed.

When creating a tracker, tracking_on is initialized to 0 and the
trackers should check this value before doing any activity. Changing the
value to 1 enables the tracker, any other value can be handled by the
trackers as they. Changing from any value > 0 to 0 clears the state HT.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>