Check for a maximum number of iterations on duplicate keys
With a too permissive configuration, the rt tracker might end up storing
an enormous amount of duplicate keys. Instead of risking to crash the
system, we can now limit the maximum number of iterations allowed in the
probes (default: 10, configurable through debugfs nr_duplicates_max).
When this value is reached, the tracker is disabled since it cannot
produce any more valid results. We cannot print any warning from this
context, so the only way to know it happened is when unloading the
tracker with the "Tracker aborted" kernel message.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>