Rename switch_out_blocked to output_switch_out_blocked
The switch_out_blocked option in the rt tracker is what prevents endless
chains of events and we cannot have smart enough filters to prevent
that, so disabling this option is not really a good idea. But sometimes
we don't want to see the switch_out_blocked events in the output trace,
so we rename this option to output_switch_out_blocked, that way the user
can disable the callback without disabling the feature.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>