Beginning of time-to-first-byte tracker
This tracker mesures the delay between the return of the accept system
call and the entry of the first write/writev system call related to the
same inode (so it works with FD inherited to child processes, dup,
worker threads, etc). This is a work in progress, it still needs
filtering (procname and PID) as well as a tracepoint for the output, but
it works.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>