Default instance name
Instead of putting all the configuration files for a tracker inside
/sys/kernel/debug/latency/$TRACKER, put them in
/sys/kernel/debug/latency/$TRACKER/default.

"default" is the name of the instance. For now it is hardcoded and only
one instance can be created through the CLI, but it will allow to have
multiple instances in the future without breaking the current usage.

The main limitation right now is that debugfs does not allow to hook on
mkdir/rmdir system calls, so adding the support for multiple instances
is not trivial, but the underlying code of the tracker is ready.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>