Control the memory allocation from debugfs
It is now possible to configure the memory allocated to a tracker by
writing in the debugfs "alloc_size" file before starting the tracker.

The memory is only allocated when writing 1 to the new "allocate"
debugfs file or when starting the tracker (echo 1 > tracking_on). For
now, the memory can only be freed by removing the tracker module.

Once the memory is allocated, the alloc_size cannot be changed anymore.

We also expose the max_resize value to debugfs.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>