Flag to follow the longest delays
The new worst_threshold debugfs flag when set to 1 allows the trackers
to automatically bump the threshold everytime a measured latency is
higher than the current threshold.

This allows to only keep track of the longest delays and could be used
in production for days or months to only get those abnormal latencies.

The flag can be reset to 0 at run-time and the current threshold can
also be changed at run-time.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>